Adds help and abort in case of failed build in build script

+++ build_and_deploy.sh
@@ -6,9 +6,16 @@ target_directory="./public"
 # Parse command line flags
 while [ $# -gt 0 ]; do
     case "$1" in
-        --target-directory)
+        --target-directory | -t)
             shift
             target_directory="$1"
+            ;;
+        --help | -h)
+            echo "Usage: $0 [OPTIONS]"
+            echo "Options:"
+            echo "  --target-directory, -t  Specify the target directory"
+            echo "  --help, -h              Display this help message"
+            exit 0
             ;;
         *)
             # Unknown flag
@@ -37,10 +44,14 @@ case "$response" in
         ;;
 esac
 
-sudo docker build -t hugo-release -f Containerfile .
-# create target directory if not exist
-mkdir -p $target_directory
-# clear target directory
-rm -rf $target_directory/*
-# copy build to target directory
-sudo docker run --rm hugo-release tar -cf - public | tar -xvf - -C $target_directory --strip-components=1
+if sudo docker build -t hugo-release -f Containerfile .; then
+    # create target directory if not exist
+    mkdir -p $target_directory
+    # clear target directory
+    rm -r $target_directory/*
+    # copy build to target directory
+    sudo docker run --rm hugo-release tar -cf - public | tar -xvf - -C $target_directory --strip-components=1
+else
+    echo "Docker build failed, aborting..."
+    exit 1
+fi
